PartsTracker/requirements.txt
# Verified working requirements - tested with all module imports
# Generated from clean venv on 2026-01-20

amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
azure-core==1.38.0
azure-storage-blob==12.28.0
billiard==4.2.4
cascadio==0.0.13
celery==5.6.2
# 3D model processing (mesh optimization)
fast-simplification>=0.1.7
trimesh>=4.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
cryptography==44.0.2
dj-rest-auth==7.0.2
Django==5.1.6
django-allauth==65.5.0
django-auditlog==3.1.2
django-browser-reload==1.18.0
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-extensions==3.2.3
django-filter==25.1
django-formtools==2.5.1
django-redis==6.0.0
django-storages==1.14.6
django-tailwind==3.8.0
django-timezone-field==7.2.1
django-widget-tweaks==1.5.0
django_celery_results==2.6.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
Faker==37.1.0
greenlet==3.3.0
idna==3.11
inflection==0.5.1
isodate==0.7.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
numpy==2.4.1
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pgvector==0.4.2
pillow==12.1.1
playwright==1.57.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.23
pyee==13.0.0
PyJWT==2.10.1
PyPDF2==3.0.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uuid-utils==0.14.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
